single/requirements.txt
aiohttp==3.9.1
aiomysql==0.2.0
tqdm==4.66.1
cryptography==41.0.7
pandas>=1.3.0
openpyxl>=3.0.0 